refactored, putting all validation methods in DataValidator
