+++ src/main/java/employee_csv_project/controller/csv_controller/data_transformation/SalaryConverter.java
@@ -1,0 +1,11 @@
+package employee_csv_project.controller.csv_controller.data_transformation;
+
+public class SalaryConverter {
+
+    public static float convertSalaryToFloat(String givenSalary) {
+        float salary = Float.parseFloat(givenSalary);
+        return salary;
+    }
+
+
+}
