--- src/main/java/employee_csv_project/controller/csv_controller/data_validation/EmailChecker.java
@@ -1,8 +1,0 @@
-package employee_csv_project.controller.csv_controller.data_validation;
-
-public class EmailChecker {
-
-    public static boolean validateEmail(String emailToCheck){
-        return emailToCheck.matches("^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+$");
-    }
-}
