+++ src/main/java/employee_csv_project/controller/csv_controller/data_validation/EmployeeCheck.java
@@ -1,7 +1,6 @@
 package employee_csv_project.controller.csv_controller.data_validation;
 
 import employee_csv_project.controller.csv_controller.data_transformation.DataCleaner;
-import employee_csv_project.controller.csv_controller.data_transformation.SalaryValidater;
 
 import java.time.LocalDate;
 
@@ -14,11 +13,11 @@ public class EmployeeCheck {
         boolean checkMiddleInit = DataValidator.validateMiddleInit(employee[3]);
         //boolean checkLastName;
         boolean checkGender = DataValidator.validateGender(employee[5]);;
-        boolean checkEMail = EmailChecker.validateEmail(employee[6]);
+        boolean checkEMail = DataValidator.validateEmail(employee[6]);
         LocalDate dateOfBirth = DataCleaner.formatDate(employee[7]);
         LocalDate dateOfJoining = DataCleaner.formatDate(employee[8]);;
         boolean checkDobBeforeDateOfJoining = DataValidator.validateDates(dateOfBirth,dateOfJoining);
-        boolean checkSalary = SalaryValidater.formatSalary(employee[9]);
+        boolean checkSalary = DataValidator.formatSalary(employee[9]);
 
         if (checkMiddleInit && checkGender && checkEMail && checkDobBeforeDateOfJoining && checkSalary){
             return true;
